refactor: relocate game-screen and dialogue fixtures outside support tree (#4117)
Move game_screen_test_support, province_draft_orders_test_support,
yarn_test_fixtures, and game_map_area_state_logic_test_support from
app/test/support/ to app/test/ with shrink-only re-export stubs.
Ratchet appTestSupportLocCeiling 2833 → 2261 (−572 LOC).

Refs #4117
